Add DebouncedInput, using HOC
Note that there's a level of indirection within DebouncedComponent.
This is to support the idea that I may one day use the Render Props
pattern, instead of an HOC. I believe that would offer better
expressiveness and much less code, but I can't figure out how to
combine the debouncing with flushing the pending actions on blur.